refactor : friend 도메인 member 패키지(모듈) 내부로 이동
