+++ src/main/java/univ/earthbreaker/namu/database/core/member/FriendJpaEntity.java
