+++ src/main/java/univ/earthbreaker/namu/database/core/member/FriendJpaRepository.java
