+++ src/main/java/univ/earthbreaker/namu/database/core/member/FriendRepositoryAdapter.java
